feat: UserBuilder 클래스로 분리
User객체를 생성하는 다양한 방법을 지원하기 위해 User에 있던 빌더를 클래스로 추출
