+++ src/main/java/db/Database.java
@@ -1,6 +1,6 @@
 package db;
 
-import model.User;
+import model.User.User;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
